Route testing

+++ pawsome-walks/helper-functions/ownerFavouriteWalks-table/deleteOwnerFavouriteWalksByBothIds.js
@@ -1,10 +1,13 @@
 import { pool } from "../../index.js";
 
-export default async function deleteOwnerFavouriteWalksByOwnerId(
+export default async function deleteOwnerFavouriteWalksByBothIds(
   ownerId,
   walkId
 ) {
   try {
+    if (!ownerId || !walkId) {
+      throw new Error("ownerId and walkId are required.");
+    }
     const deleteOwnerFavouriteWalks =
       "DELETE FROM ownerFavouriteWalks WHERE ownerId = $1 AND walkId = $2 RETURNING *";
     const result = await pool.query(deleteOwnerFavouriteWalks, [
